# src/Api.Client/Generator/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AErmilov.NumbersIntoWords.Api",
    "version": "1.0"
  },
  "paths": {
    "/numbers/{number}/words": {
      "get": {
        "tags": [
          "NumbersIntoWords"
        ],
        "operationId": "ConvertNumberIntoWords",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number",
              "format": "decimal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NumberAsWordsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NumberAsWordsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NumberAsWordsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "NumberAsWordsResponse": {
        "type": "object",
        "properties": {
          "number": {
            "type": "number",
            "format": "decimal"
          },
          "words": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}